added buffers -VertexBuffer, ElementBuffer and also VertexArray with a BufferLayout object, which can take in custom structs-

+++ LearnOpenGL/src/Timer.h
@@ -11,7 +11,7 @@ namespace utilities
 
 	public:
 		Timer() { }
-		float getDeltaTime();
+		long long getDeltaTime();
 		void resetTimer();
 
 	private:
